MC-20508: Api-functional tests for Product prices including FPT
- added FPT schema implementation
- deprecated price adjustments in schema
- added test coverage for product with FPT attributes

+++ dev/tests/integration/testsuite/Magento/Weee/_files/fixed_product_attribute_rollback.php
@@ -3,12 +3,24 @@
  * Copyright © Magento, Inc. All rights reserved.
  * See COPYING.txt for license details.
  */
-
 declare(strict_types=1);
 
-$objectManager = \Magento\TestFramework\Helper\Bootstrap::getObjectManager();
+use Magento\TestFramework\Helper\Bootstrap;
 
-/* @var EavAttribute $attribute */
-$attribute = $objectManager->get(\Magento\Eav\Model\Entity\Attribute::class);
-$attribute->loadByCode(\Magento\Catalog\Model\Product::ENTITY, 'fixed_product_attribute');
-$attribute->delete();
+$objectManager = Bootstrap::getObjectManager();
+/** @var \Magento\Framework\Registry $registry */
+$registry = $objectManager->get(\Magento\Framework\Registry::class);
+
+$registry->unregister('isSecureArea');
+$registry->register('isSecureArea', true);
+
+/** @var \Magento\Catalog\Model\ResourceModel\Eav\Attribute $attribute */
+$attribute = \Magento\TestFramework\Helper\Bootstrap::getObjectManager()
+    ->create(\Magento\Catalog\Model\ResourceModel\Eav\Attribute::class);
+$attribute->load('fixed_product_attribute', 'attribute_code');
+if ($attribute->getId()) {
+    $attribute->delete();
+}
+
+$registry->unregister('isSecureArea');
+$registry->register('isSecureArea', false);
